Create target for test application, split it up
