+++ Sources/TestApplication/TodoDomain.swift
@@ -1,0 +1,7 @@
+// Pure models
+
+struct Todo {
+    var id: String?
+    let title: String
+    let done: Bool
+}
