+++ Sources/TestApplication/TodoMapping.swift
@@ -1,0 +1,6 @@
+// Mapping domain to and from intermediate representations
+
+import Quark
+
+// Default implementation through magic read/write reflection
+extension Todo : ResourceConvertible {}
