+++ Sources/TestApplication/main.swift
@@ -1,0 +1,9 @@
+// Bring everything together
+
+import Quark
+
+let store = TodoInMemoryStore()
+let app = TodoApplication(store: store)
+let router = TodoRouter(app: app)
+
+try Server(router).start()
